+++ changes.md
@@ -6,6 +6,7 @@
 * can.CanTower added render_rails parameter.
   * bp_rails can be set to None.
 * Changed build order for CanRail so that the cut operation happens after supports are added. 
+  * calculate cut rail height
 
 ## 2.4.1
 * documentation clean up